refactor: improve excluded coverage files list
